Move code to either abstract Matrix of concrete implementing class
Closes #42
