+++ src/main/java/mr/cookie/matrix/model/SingleThreadMatrix.java
@@ -3,6 +3,9 @@ package mr.cookie.matrix.model;
 import mr.cookie.matrix.random.Random;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.List;
+import java.util.stream.IntStream;
+
 /**
  * A single thread implementations of a Matrix class that represents a matrix.
  *
@@ -21,6 +24,36 @@ public final class SingleThreadMatrix extends Matrix {
      */
     public SingleThreadMatrix(int rowSize, int columnSize, @NotNull int... elements) {
         super(rowSize, columnSize, elements);
+    }
+
+    /**
+     * Returns a new matrix that is a product of multiplying two input matrices. An input matrix needs to have
+     * the same row count as the other matrix has column count (and vice versa). If not then
+     * {@link IllegalArgumentException} will be thrown.
+     *
+     * @param m1 first matrix that will be used during multiplication
+     * @param m2 second matrix that will be used during multiplication
+     * @return a new matrix that is a product of multiplying two matrices
+     */
+    public static @NotNull Matrix multiply(@NotNull Matrix m1, @NotNull Matrix m2) {
+        verifyRowAndColumnCountsForMultiplication(m1, m2);
+
+        final int rowSize = m1.getRowSize();
+        int[] numbers = new int[rowSize * rowSize];
+        int index = 0;
+
+        for (int c = 0; c < rowSize; c++) {
+            for (int r = 0; r < rowSize; r++) {
+                List<Integer> row = m1.getRow(c);
+                List<Integer> column = m2.getColumn(r);
+
+                numbers[index++] = IntStream.range(0, row.size())
+                        .map(i -> row.get(i) * column.get(i))
+                        .sum();
+            }
+        }
+
+        return new SingleThreadMatrix(rowSize, rowSize, numbers);
     }
 
     /**
